+++ tests/accelerators/test_accelerator_connector.py
@@ -632,23 +632,23 @@ def test_accelerator_ddp_for_cpu(tmpdir):
     assert isinstance(trainer.training_type_plugin, DDPPlugin)
 
 
-def test_exception_when_accelerator_strategy_used_with_distributed_backend():
+def test_exception_when_strategy_used_with_distributed_backend():
     with pytest.raises(MisconfigurationException, match="but have also passed"):
-        Trainer(distributed_backend="ddp_cpu", accelerator_strategy="ddp_spawn")
+        Trainer(distributed_backend="ddp_cpu", strategy="ddp_spawn")
 
 
-def test_exception_when_accelerator_strategy_used_with_accelerator():
+def test_exception_when_strategy_used_with_accelerator():
     with pytest.raises(MisconfigurationException, match="but have also passed"):
-        Trainer(accelerator="ddp", accelerator_strategy="ddp_spawn")
+        Trainer(accelerator="ddp", strategy="ddp_spawn")
 
 
-def test_exception_when_accelerator_strategy_used_with_plugins():
+def test_exception_when_strategy_used_with_plugins():
     with pytest.raises(MisconfigurationException, match="only specify one training type plugin, but you have passed"):
-        Trainer(plugins="ddp_find_unused_parameters_false", accelerator_strategy="ddp_spawn")
+        Trainer(plugins="ddp_find_unused_parameters_false", strategy="ddp_spawn")
 
 
 @pytest.mark.parametrize(
-    ["accelerator_strategy", "plugin"],
+    ["strategy", "plugin"],
     [
         ("ddp_spawn", DDPSpawnPlugin),
         ("ddp_spawn_find_unused_parameters_false", DDPSpawnPlugin),
@@ -656,20 +656,20 @@ def test_exception_when_accelerator_strategy_used_with_plugins():
         ("ddp_find_unused_parameters_false", DDPPlugin),
     ],
 )
-def test_accelerator_strategy_choice_cpu_str(tmpdir, accelerator_strategy, plugin):
-    trainer = Trainer(accelerator_strategy=accelerator_strategy, accelerator="cpu", devices=2)
+def test_strategy_choice_cpu_str(tmpdir, strategy, plugin):
+    trainer = Trainer(strategy=strategy, accelerator="cpu", devices=2)
     assert isinstance(trainer.training_type_plugin, plugin)
 
 
 @pytest.mark.parametrize("plugin", [DDPSpawnPlugin, DDPPlugin])
-def test_accelerator_strategy_choice_cpu_plugin(tmpdir, plugin):
-    trainer = Trainer(accelerator_strategy=plugin(), accelerator="cpu", devices=2)
+def test_strategy_choice_cpu_plugin(tmpdir, plugin):
+    trainer = Trainer(strategy=plugin(), accelerator="cpu", devices=2)
     assert isinstance(trainer.training_type_plugin, plugin)
 
 
 @RunIf(min_gpus=2)
 @pytest.mark.parametrize(
-    ["accelerator_strategy", "plugin"],
+    ["strategy", "plugin"],
     [
         ("ddp_spawn", DDPSpawnPlugin),
         ("ddp_spawn_find_unused_parameters_false", DDPSpawnPlugin),
@@ -682,15 +682,15 @@ def test_accelerator_strategy_choice_cpu_plugin(tmpdir, plugin):
         pytest.param("deepspeed", DeepSpeedPlugin, marks=RunIf(deepspeed=True)),
     ],
 )
-def test_accelerator_strategy_choice_gpu_str(tmpdir, accelerator_strategy, plugin):
-    trainer = Trainer(accelerator_strategy=accelerator_strategy, accelerator="gpu", devices=2)
+def test_strategy_choice_gpu_str(tmpdir, strategy, plugin):
+    trainer = Trainer(strategy=strategy, accelerator="gpu", devices=2)
     assert isinstance(trainer.training_type_plugin, plugin)
 
 
 @RunIf(min_gpus=2)
 @pytest.mark.parametrize("plugin", [DDPSpawnPlugin, DDPPlugin])
-def test_accelerator_strategy_choice_gpu_plugin(tmpdir, plugin):
-    trainer = Trainer(accelerator_strategy=plugin(), accelerator="gpu", devices=2)
+def test_strategy_choice_gpu_plugin(tmpdir, plugin):
+    trainer = Trainer(strategy=plugin(), accelerator="gpu", devices=2)
     assert isinstance(trainer.training_type_plugin, plugin)
 
 
@@ -698,7 +698,7 @@ def test_accelerator_strategy_choice_gpu_plugin(tmpdir, plugin):
 @pytest.mark.parametrize("plugin", [DDPSpawnPlugin, DDPPlugin])
 def test_device_type_when_training_plugin_gpu_passed(tmpdir, plugin):
 
-    trainer = Trainer(accelerator_strategy=plugin(), gpus=2)
+    trainer = Trainer(strategy=plugin(), gpus=2)
     assert isinstance(trainer.training_type_plugin, plugin)
     assert trainer._device_type == DeviceType.GPU
     assert isinstance(trainer.accelerator, GPUAccelerator)