+++ tests/accelerators/test_tpu_backend.py
@@ -230,15 +230,15 @@ def test_ddp_cpu_not_supported_on_tpus():
 
 
 @RunIf(tpu=True)
-@pytest.mark.parametrize("accelerator_strategy", ["tpu_spawn", "tpu_spawn_debug"])
-def test_accelerator_strategy_choice_tpu_str(tmpdir, accelerator_strategy):
-    trainer = Trainer(accelerator_strategy=accelerator_strategy, accelerator="tpu", devices=8)
+@pytest.mark.parametrize("strategy", ["tpu_spawn", "tpu_spawn_debug"])
+def test_strategy_choice_tpu_str(tmpdir, strategy):
+    trainer = Trainer(strategy=strategy, accelerator="tpu", devices=8)
     assert isinstance(trainer.training_type_plugin, TPUSpawnPlugin)
 
 
 @RunIf(tpu=True)
-def test_accelerator_strategy_choice_tpu_plugin(tmpdir):
-    trainer = Trainer(accelerator_strategy=TPUSpawnPlugin(), accelerator="tpu", devices=8)
+def test_strategy_choice_tpu_plugin(tmpdir):
+    trainer = Trainer(strategy=TPUSpawnPlugin(), accelerator="tpu", devices=8)
     assert isinstance(trainer.training_type_plugin, TPUSpawnPlugin)